tests/testpkcs11.sh: run pkcs11 tests in the year of 2014
The tests/testpkcs11.sh has been renamed to test/testpkcs11_orig.sh,
which is now called by the new tests/testpkcs11.sh with the datefudge,
that allows the pkcs11 tests to be run in the year of 2014.
There is a certain pkcs11 CA cert, which is pregenerated in 2013 and
it has been expired in 2020, that caused failing the pkcs11 tests, so
it needs to be run in a year between 2013 and 2020.

This patch should be reverted after gnutls developers has made better
time invariant tests or regenerated that certain CA cert.

References: jira PAM-13677
References: jira PAM-13257